modularize comment.js
